[menu&events branch] - Ported research dialog - Ported upgrades menu - Fixed voice sounds getting cut by other sound effects - Fixed some warnings (float vs. double)
git-svn-id: svn://maxr.org/MAX/branches/menu_event_refactoring@3564 c3bd5e88-ca0b-48cf-b972-da51c77c4c1b

+++ gui/game/temp/unitdrawingengine.cpp
@@ -108,7 +108,7 @@ void cUnitDrawingEngine::setDrawColor (bool drawColor)
 }
 
 //--------------------------------------------------------------------------
-void cUnitDrawingEngine::drawUnit (const cBuilding& building, SDL_Rect destination, double zoomFactor, const cUnitSelection* unitSelection, const cPlayer* player)
+void cUnitDrawingEngine::drawUnit (const cBuilding& building, SDL_Rect destination, float zoomFactor, const cUnitSelection* unitSelection, const cPlayer* player)
 {
 	const auto animationFlags = animationTimer->getAnimationFlags ();
 
@@ -262,7 +262,7 @@ void cUnitDrawingEngine::drawUnit (const cBuilding& building, SDL_Rect destinati
 }
 
 //--------------------------------------------------------------------------
-void cUnitDrawingEngine::drawUnit (const cVehicle& vehicle, SDL_Rect destination, double zoomFactor, const cMap& map, const cUnitSelection* unitSelection, const cPlayer* player)
+void cUnitDrawingEngine::drawUnit (const cVehicle& vehicle, SDL_Rect destination, float zoomFactor, const cMap& map, const cUnitSelection* unitSelection, const cPlayer* player)
 {
 	const auto animationFlags = animationTimer->getAnimationFlags ();
